fix buttons in newAdminProfile
