+++ client/src/Components/Profile/NewAdminProfile/Button/Button.module.css
@@ -4,6 +4,7 @@
     padding-top: 0.5rem;
     padding-bottom: 0.5rem;
     padding-left: 7%;
+    position: relative;
     text-align: start;
     font-family: 'Roboto';
     font-style: normal;
