+++ client/src/Components/Profile/NewAdminProfile/SubItems/Subitems.jsx
@@ -7,28 +7,28 @@ export const SubItems = ({ children, onClick, active, childrens, value }) => {
   const [show, setShow] = useState(false);
 
   return (
-    <div className={s.containerBtn}>
-      <div className={s.btnAndBox}>
-        <Button onClick={() => setShow(!show)} active={active} value={value}>
-          {children}
-          <Box show={show}/>
-        </Button>
-      </div>
-
-      {show &&
-        childrens.map(({ value, table, childrens }) => {
-          return (
-            <Button
-              onClick={() => onClick(table, value)}
-              active={active}
-              value={value}
-              table={table}
-              childrens={childrens}
-            >
-              {value}
-            </Button>
-          );
-        })}
+    <div className={`${s.containerBtn} ${show ? s.active : ""}`}>
+      <Button onClick={() => setShow(!show)} active={active} value={value}>
+        {children}
+        <Box show={show} onClick={() => setShow(!show)} />
+      </Button>
+      {show && (
+        <section>
+          {childrens.map(({ value, table, childrens }) => {
+            return (
+              <Button
+                onClick={() => onClick(table, value)}
+                active={active}
+                value={value}
+                table={table}
+                childrens={childrens}
+              >
+                {value}
+              </Button>
+            );
+          })}
+        </section>
+      )}
     </div>
   );
 };
