+++ client/src/Components/Profile/NewAdminProfile/SubItems/Subitems.module.css
@@ -2,11 +2,11 @@
     width: 100%;
     overflow: hidden;
 }
-.containerBtn button:not(:nth-child(1)){
-    margin-left: 10%;
+
+.active{
+    overflow: auto;
 }
 
-.btnAndBox{
-    display: flex;
-    position: relative;
+.containerBtn section{
+    margin-left: 10%;
 }
